+++ src/views/ConfigView.vue
@@ -1,4 +1,5 @@
 <script setup>
+import { resetToDefaultAccount } from '../const';
 import { watch } from 'vue';
 import { useRouter } from 'vue-router';
 import { useI18n } from 'vue-i18n';
@@ -10,6 +11,7 @@ const router = useRouter();
 const stored_locale = useStorage('sake-locale', '');
 const { t, locale } = useI18n({ useScope: 'global' });
 const colorMode = useStorage('sake-color-mode');
+const resetToDefault = useStorage('sake-reset-to-default', resetToDefaultAccount);
 const allowedColorModes = [
 	'light',
 	'dark',
@@ -44,6 +46,12 @@ watch(colorMode, async (newColorMode) => {
 			<select class="form-select" id="app-color-mode" v-model="colorMode">
 				<option v-for="mode in allowedColorModes" :key="mode" :value="mode">{{ $t(`config.${mode}Theme`) }}</option>
 			</select>
+		</div>
+		<div class="mb-3">
+			<div class="form-check form-switch">
+				<input class="form-check-input" type="checkbox" role="switch" id="app-reset-to-default" v-model="resetToDefault">
+				<label class="form-check-label" for="app-reset-to-default">{{ $t("config.resetToDefault") }}</label>
+			</div>
 		</div>
 
 		<ButtonGroupComponent>